A bunch of commented out shit
I've gone and maybe done some of the shit for 1a. Unfortunately retargetting isn't working for me so I can't test any of my code due to 679 errors but I've marked it all with "Alex" in case you want to control-F it. I hope to work on this more over the weekend when not doing PPD or FOCA. ='(